Quick Test on getting working some stuff
It's just a quicker way of getting a random page from tcrf.net (go and
check it out, it's awesome). Also found how to handle strange windows ui
calls, like ALT + TAB, which can't be called by the class Robot by
default, but with a system shortcut, which is a pretty neat trick.